task1

+++ 0x17-doubly_linked_lists/0-print_dlistint.c
@@ -2,13 +2,13 @@
 
 /**
 * print_dlistint - Function that print all element on a list
-* @dlistint_t: the list to be printed
+* @h: the list to be printed
 * Return: number of nodes
 */
 
 size_t print_dlistint(const dlistint_t *h)
 {
-	int n = 0;
+	size_t n = 0;
 
 	while (h != NULL)
 	{
